Return text format wt user profile (#66)
