+++ core/axbot-server/src/main/java/com/github/axiangcoding/axbot/server/service/entity/SendTextWtProfile.java
@@ -1,0 +1,9 @@
+package com.github.axiangcoding.axbot.server.service.entity;
+
+import lombok.Data;
+
+@Data
+public class SendTextWtProfile {
+    String userId;
+    String text;
+}
